fix cajaServer parameter in Playground http://codereview.appspot.com/4601042
In a previous commit, I changed cajaServer from '/' to ''
to make the Playground work when installed other than at /
However, caja.configure treats '' the same as unspecified,
so cajaServer will be caja.appspot.com, rather than the
local test cajoler.

This change replaces '' with '.', which will be the local
cajoler.

R=jasvir@gmail.com, ihab.awad@gmail.com