Handling hyphens within comments gracefully <unassigned>
In the renderUnsafe mode, we currently disallow comments that begin with or end
with '-' because of the Html 5 and XML 1 specs.

This causes an IllegalStateException with the message "XML comment unrenderable
because it ends with '-'" to be thrown when parsing some documents:

However, browsers do handle these gracefully and allow these extra hyphens, as
can be seen in the following discussions:
http://www.w3.org/TR/html4/intro/sgmltut.html#h-3.2.4
http://marcorogers.com/blog/08-02-2009/strange-firefox-rendering-bug

This patch modifies comments that begin or end with hyphens or contain
consecutive hyphens, to make them spec-compliant.

Patch from Anupama. Thanks!

R=jasvir@gmail.com