Fix bug around writability of inherited properties. http://codereview.appspot.com/4625075
isWritable() was missing a test for the case when the x property is
marked writable but not yet fastpathed and an object o inherits x.
It was returning false when it should have returned true.

Also forced the result of an assignment to equal the rhs, not
the result of the setter.

R=erights@google.com