Add pure ES5+client-side-rewriting support to caja.js. http://codereview.appspot.com/4852051
Add a mode which creates taming frames and guest frames using SES instead of ES5/3, and uses client-side facilities such as HtmlEmitter rather than the cajoling service. Adapt the test suite to run selected browser tests under both SES and ES5/3. The test suite therefore now requires an SES-capable browser.

To restore compatibility with SES on Firefox, Domado and SES are changed to throw out the fragile ses_ignoreBug_propertyWillAppearAsOwn kludge in favor of simply letting the affected properties be non-enumerable.

Also added a SES-compatible version of Q (ses-promise.js, after cajita-promise.js) for use by caja.js, and fixed all overlength lines in domado.js.

Caveats:

There is not yet auto-detection of browser ES5 support.

Guest code is run under SES, so it sees frozen primordials in the guest frame, which is not quite what we want. Fixing this is pending MarkM coming up with a "CES" system which is SES without the freezing.

Scripts and stylesheets embedded in or linked from HTML do not take effect.

R=jasvir@gmail.com, ihab.awad@gmail.com