First pass at support for SCRIPT and STYLE elements in document.write content. http://codereview.appspot.com/4909044
Under SES, we want to be able to skip any server side transform of
HTML and CSS.  A container should be able to get a handle to the tame
document.write and simply write the gadget HTML.

This is some initial plumbing for the special handling needed for
SCRIPT and STYLE handling.

It delegates SCRIPT handling to cajaVM.eval and just has a stub for
CSS handling.

Open questions:
(1) How do we reroute through a rewriting proxy for ES5/3?
(2) Is the cajaVM.eval call appropriate?

Still TODO:
(1) Client side CSS rewriting.
(2) Proper SES testing.
(3) Sharing of tag stack between the HTML sanitizer and HTML emitter
    which probably requires a resumable tag-stack for document.written content.

R=jasvir@gmail.com, erights@google.com