faster es53 freeze http://codereview.appspot.com/5484074
This change reduces runtime of a minimal test gadget
from 2800ms to 1800ms in Chrome Beta (16.0),

Basically, much of the cost of es53+domado startup is freeze,
and much of the cost of freeze is iterating over object properties,
which includes inherited properties which have to be ignored.

This change tries to use a native Object.keys function to do
a fast initial filtering of the properties.

R=metaweta@gmail.com