fix es53 interference with chrome debugger http://codereview.appspot.com/5494047
the debugger in chrome expects to call Object.getOwnPropertyDescriptor
on arbitrary objects from any frame, but es53 overrides gOPD with a
version that only works for objects that inherit from es53's Object.

This change virtualizes gOPD to avoid interfering with the debugger.

Previously we were virtualizing Object.getOwnPropertyNames for
the same reason.

R=metaweta@gmail.com