Reworked client side property verification and normalization http://codereview.appspot.com/5369112
This replaces the regular expression matching scheme for CSS property values
with one that is token order independent but which allows for faster matching
and better normalization.

CssPropertyPatterns now generates a css-defs.js that is in a different format.
domado.js and other clients have been updated to use cssmatcher.js which uses
css-defs.js to verify and normalize CSS tokens against a property schema
element.

R=jasvir@gmail.com