precajoler fixes http://codereview.appspot.com/5489121
1. PrecajoleRewriteStage fails to substitute more than one precajoled
script, due to incorrect looping over a dom tree being modified.

This change fixes that, and adds some test cases for that.

2. The ant PrecajoleTask cajoles js files specified in the build.xml
file, which means that changing build.xml forces everything to be
re-cajoled, which is annoying enough now, and will become more annoying
when we add more things to precjaole.

This change moves the precajole list to a separate xml file, so that
only changes to that xml file will force everything to be re-cajoled.

R=jasvir@gmail.com