Fixes several failing jQuery tests http://codereview.appspot.com/5654047
The lexCss() function requires a string as input but some callers weren't
casting first.  I added a cast at the start of the function.

If you wrap the onreadystatechange property of XHR in an untame(), then
when the feral native DOM layer tries to call the event callback, it will
try to tame the |event| object passed in, which is neither possible 
(it's an unsupported weirdo object) nor necessary,
so the correct thing to do is just to pass a plain function.

We do not tame xhr.responseXML, but jquery tests crash if it does not
contain at least documentElement.cloneNode.  I added a kludge that allows
the jQuery tests to continue to run and a message that Caja doesn't
support it.

attachDocumentStub() was renamed to attachDocument() a while back; this
changes the error message to match.

In one situation in jquery that I haven't been able to replicate outside
it, traversing the parentNode chain can lead to a host page DOM node that
has not yet had the caja properties attached so that cajoled Domado can
use them.  This change makes sure the properties are there.

Before this change, the HTML element violated the spec that says the
parentNode of the element should be the document, so this change
fixes that.

R=ihab.awad