Test document.all. http://codereview.appspot.com/5698077
document.all is native object on IE such that document.all[x] is similar to
document.getElementById(x).

It is worth separate testing because Firefox supports it, but treats it
specially so that:

   typeof document.all === "undefined"
   !!document.all

so that feature detection code will take the no document.all path, but code
that does not feature detect will continue to work in quirks mode:

   document.all.x

is not a TypeError.  It is in the square free shell which is in standards
compliance mode.

http://weblogs.mozillazine.org/asa/archives/007118.html explains:
"""
December 17, 2004
mozilla 1.7.5 released

...

The new undetectable document.all support, and support for exposing elements by their ID in the global scope for greater IE compatibility when viewing pages that don't request standards compliant behaviour is documented in bugs 248549 and 256932.
"""

R=felix8a@gmail.com