browser test fixes, mainly for IE http://codereview.appspot.com/5733044
This is mostly about support for manually running the browser tests
on IE.

1. Many tests have localhost:8000 hardcoded, which works when testdriver
and browser are on the same machine, but doesn't work when they're not.

2. Add <!doctype html> to the browser test files so IE is automatically
in standards mode not quirks mode.

3. In standards mode, es53-test-container-overflow fails in firefox
because document.body.clientHeight is shrinkwrapped: it's the size of
the document content, not the size of the window.  document.documentElement
has the clientHeight we need.

4. "ant brtest" tries to antcall "runtests" twice, but if the first antcall
fails, the second doesn't get run.  The first antcall can fail if the
test run succeeds, but ant-reports has failing test results from a
previous run.  The simplest fix is to have "ant brtest" call runtests
only once, by renaming both browser tests to have a common name suffix.

R=metaweta@gmail.com