Fixes a bug around assignment and no setters and an inner hull breach. http://codereview.appspot.com/5873043
Makes ES5/3 conform to the ES5 spec and throw when trying to assign
a value to an accessor property with no setter.

Toxic functions could be 'laundered' by using call, bind, or apply,
so an outer hull breach could be escalated to an inner hull breach.

I discovered the inner hull breach when I noticed that the wrong 
exceptions were being thrown in ES5Rewriter; since jsUnit doesn't
provide a way to check the kind of exception, I added 
   assertThrowsMsg(fn, msg)
that fails unless an exception ex is thrown and msg is a substring of
ex.message.

R=felix8a