Rework Server side CSS passes so CLASS and ID namespacing occurs in CSSRewriter http://codereview.appspot.com/5975048
This changes CssDynamicExpressionRewriter and CssRewriter so now CssRewriter
is responsible for namespacing IDs and CLASSes in CSS selectors.

It updates the tests to include the rewritten selectors.

I would prefer to implement the CSS selector rewriter to
sanitizecss.js in another CL and reenable the JS side CSS tests that
are now marked @FailureIsAnOption to keep this CL reasonably small.

R=jasvir@gmail.com