Support unicode characters and \u escapes in ES5 mode http://codereview.appspot.com/5992072
Adds support for unicode characters and \u escaped characters in ES5 mode.
This is needed for jQuery which uses \u escapes in source in regexs.

The change works by rewriting all [^\u0000-\u007f] characters with their \u escaped version.  It then detects all
identifiers in atLeastFreevars by including \u escapes in the regular expression.  This means some identifiers which
aren't valid such as a\u000ab are included but that continues to meet the conservative criteria of atLeastFreeVars.

R=felix8a@gmail.com