remove a['x'] optimization http://codereview.appspot.com/6066048
Caja's js renderer always tries to rewrite (a['x']) as (a.x),
which is not semantically equivalent when using closure compiler.

a['x'] is rare usage except when targeting closure compiler.
the a.x form is not any faster in browsers.
the rewrite saves only 3 bytes per use.

the optimization is not much of a win, so I'm just removing it.

R=jasvir@gmail.com