allow caller to supply both cajoled and uncajoled code http://codereview.appspot.com/6035056
In some cases, a server like shindig has both the cajoled and uncajoled
version of some guest code, and on the server side we don't know if
the browser will support ses or not.

So we want to send to the browser code that will create a guest
frame and then do
   frame.code(...).cajoled(...).run()
so when ses is active, the code(...) will be used,
but when es53 is active, the cajoled(...) will be used.

Currently that will throw an error.  caja.js doesn't think it's
sensible to supply both uncajoled and cajoled code.

This change relaxes that restriction.

R=jasvir@gmail.com, metaweta@gmail.com