speed up junit http://codereview.appspot.com/6250057
We're running junit with fork="yes" because we want the junit vm to
have different properties than the ant vm.  However, we're spawning
a new vm for each junit test, which is unnecessary.

This adds forkmode="once" to the junit tests.

This reduces 'ant runtests' on my laptop from about 11 minutes to
about 7 minutes.

R=metaweta@gmail.com