safer html/xml rendering, phase 1: validation http://codereview.appspot.com/6274043
Currently, we have two main DOM rendering modes: HTML and XML.

We expect Caja integrators to choose the right one.  But this is a critical
security decision, because making the wrong choice allows attackers to
sneak stuff through HTML/XML misinterpretation.  And it's not particularly
obvious that this is a critical security decision.

What I'm proposing instead is that Caja always emit a safe rendering of
HTML or XML.

This patch is phase 1 of that.  It adds a framework for testing safety
and fidelity of DOM rendering.

The tests marked @FailureIsAnOption are things that I think should
be true, but are not yet true.  I'll fix them in subsequent patches,
but I wanted to get this out first and see what people think about
this plan.

R=ihab.awad@gmail.com