match arbitrary source against the precajole map http://codereview.appspot.com/6299057
Currently, precajoled modules will only be used if someone writes
a <script> tag with a uri that we recognize as something we have
precajoled.  If someone uses a version of jquery that we precajole,
but uses the wrong uri, we won't recognize it, so we'll fetch,
parse, and re-cajole it.

This patch tries to avoid that.  We can't avoid the fetch of
an unknown script uri, but once we've fetched the script we can see
if the text matches something we've precajoled, which lets us avoid
the expense of parsing and cajoling.

Scripts inlined in <script> tags will also get replaced with their
precajoled equivalent, if the text is identical.

R=metaweta@gmail.com