some fixes for html/xml rendering http://codereview.appspot.com/6308071
This is a couple steps closer toward robust html/xml rendering.

In particular

- empty tags like <div></div> used to be rendered as <div /> in xml mode,
but now it's rendered as <div></div>, which works fine as xml, and won't
be misunderstood by html parsers.

- xml cdata blocks like <div><![CDATA[&]]></div> used to be rendered
with the cdata quoting.  however, <![CDATA[]]> has no meaning in html
and might confuse html parsers, so instead that gets emitted as
<div>&amp;</div>, which is equivalent in xml and has the same
interpretation in html.

- the handling of RAWTEXT html elements like <xmp> has been cleaned up
a bit and made more robust.  this still isn't ideal.  there are still
some misinterpretation hazards that are solvable but aren't solved yet,
marked as TODO.

R=metaweta@gmail.com