relax js html parser http://codereview.appspot.com/6310049
Currently, html-sanitizer.js rejects unknown/forbidden tags in the
html parser as well as in the sanitizer.  Projects using just the
parser would like to see the unknown tags.

This patch does that.  Parsing of unknown tags is not necessarily
correct, since an unknown tag might have special parsing rules
and we treat it generically, but this has no impact on sanitizer
safety.

We keep an old html-sanitizer.js for regression testing, and this
is an incompatible change to html-sanitizer behavior, so I had to
fiddle with testing so we still have good regression coverage,
as well as coverage of the new behavior.

R=metaweta@gmail.com