Turkish i defense in java http://codereview.appspot.com/6300084
Because of the Turkish 'i', String#toLowerCase etc are unsafe to use.
We have Strings.toLowerCase etc that are safe, but they aren't used
universally.

This patch does:

- rename Strings.toLowerCase etc to Strings.lower etc, so that it's
easier to use a simple grep to find stray uses of the unsafe methods.

- add a shell script tools/codecheck.sh that complains if the unsafe
methods get used.  "ant runtests" will run codecheck.

- fix all uses of the unsafe methods.

- replace hand-coded loops in Strings.lower etc with standard calls
using Locale.ENGLISH, because it's less code to worry about.  This
behaves differently for non-ASCII characters, but we don't care as
long as it's consistent and isn't confused by Turkish 'i' issues.

R=ihab.awad@gmail.com