security fix for FlashBridge ExternalInterface.call http://codereview.appspot.com/6326053
Main thing in this patch: ExternalInterface.call is unsafe because
it basically just injects values into a javascript template that
gets eval'ed, and Flash fails to escape the values correctly.
Adobe is not planning on fixing this anytime soon, because it
might break stuff.  More details here:
http://lcamtuf.blogspot.com/2011/03/other-reason-to-beware-of.html

So I'm fixing the hole in a way that makes my skin crawl,
but I can't think of a better fix.

R=metaweta@gmail.com