remove deprecation warning for tree walkers http://codereview.appspot.com/6334059
ParseTreeNode#acceptPreOrder and #acceptPostOrder are marked
@Deprecated because they run in O(n**2) time (worst-case).
I'd like to replace all uses of them, but doing so is
non-trivial and low priority.

It's kind of annoying to see the deprecation warnings all the
time.

So this patch removes the @Deprecated annotations for now,
so an eclipse build now shows zero errors and zero warnings.

R=metaweta@gmail.com