classpath cleanup http://codereview.appspot.com/6350053
build.xml has a confusing maze of classpath definitions.
this reduces the mess.

there used to be a distinction between 'compile' classpaths
and 'run' classpaths, and the difference was that 'run' included
ant-lib while 'compile' didn't.  but that was a lie, the
'compile' classpath also included ant-lib indirectly.

I don't see any reason to maintain the distinction between
compile classpaths and run classpaths.

R=metaweta@gmail.com