force autocomplete off http://codereview.appspot.com/6343068
We force autocomplete=off in the server-side sanitizer but not in the
client-side sanitizer.  This fixes that.

I'm doing this in domado.js only, not in html-sanitizer.js, because
html-sanitizer.js doesn't currently enforce any similar constraints.
In particular, it doesn't force there to be a target= value.  But
I've added a TODO.

R=jasvir@gmail.com