loosen rawtext/cdata rendering restriction http://codereview.appspot.com/6416046
A previous change narrowed the charset we would emit when rendering
RAWTEXT contexts, because '\0' in particular can confuse html parsers.

The restriction is too narrow; some places rely on being able to
emit unicode chars in RAWTEXT, and there's no way to do encoding
in RAWTEXT.  (In the specific case of <script> nodes, it's usually
possible to rewrite the script to avoid non-ascii chars, and we
do that for the js that we generate, but this renderer is also
used to render dom trees other than the ones we generate.)

So this change loosens the charset restriction and allows everything
except '\0' chars.

R=jasvir@gmail.com