Implement document.title correctly per HTML5. http://codereview.appspot.com/6459089
The previous implementation returned the <title> node rather than the title
text, which is entirely wrong as far as I know, and also causes an incidental
error in the jQuery test suite. This implementation is according to the HTML5
spec, except insofar as our pseudo-DOM (to be replaced soon) makes that
difficult.

R=jasvir@gmail.com