known-fail for firefox 15+ freeze problem http://codereview.appspot.com/6484058
This adds some known-fail logic around the browser tests that fail due
to the firefox 15+ freeze bug.  For now known-fail is just treated
like a pass, because I'm planning on doing a more systematic method
of handling known-fails once I get testing running on a matrix of
browser versions.  This is just a quick fix to keep 'ant runtests'
running smoothly on near-future firefox versions.

R=metaweta@gmail.com