setting .length can delete non-configurable properties http://codereview.appspot.com/6495102
Setting a.length can delete non-configurable properties of a.
This is fixed in current versions of firefox and safari,
but Chrome Canary still has this bug.

So this adds a test for that in repairES5.js.
This is similar to the pop() ignores frozenness problem,
but Chrome passes the pop test and fails this test.

R=erights@gmail.com