Replace pseudo-elements with actual DOM elements. http://codereview.appspot.com/6461102
Previously, the global structure elements (<html>, <head>, <title>,
<body>) of a virtual document are provided as “pseudo-elements” which
are not backed by real DOM nodes. This has several disadvantages,
broadly that when an input is a HTML document rather than a document
fragment, very little of the information which may be stored in the HEAD
is preserved or made editable; as well, subtle code is required to
implement the pseudo-elements correctly.

Instead, we now provide the global structure elements by using real DOM
nodes but by renaming ("virtualizing") the element names (to html___,
head___, etc.). The cajoler and HtmlEmitter now output complete HTML
documents, and Domado assumes it receives a complete virtualized
document rather than body content directly in a HTML node.

To support this, html-emitter.js now implements the relevant portion of
the HTML5 tree construction algorithm.

This patch is a draft; there are a number of issues with it, and it is
being published for early review only.
* The modification to the cajoler may not be sound.
* document.write is broken, because HtmlEmitter's attach/detach is not
  aware of the new HTML5 insertion mode concept.
* Clients of html-sanitizer.js or the cajoler which intend to process
  HTML fragments, as opposed to documents, will be broken; new APIs
  need to be designed to specify the desired result.
* The element renaming is implemented in more separate places than it
  should be.

R=mikesamuel@gmail.com