whitelist event handlers uniformly http://codereview.appspot.com/6554054
currently, you can't set an onchange= listener on a div, either as an
attribute or using addEventListener, because it's not whitelisted,
because the html4 spec implies that you can't do that.

however, you actually can, and it's sensible to do so because events
bubble: it's common to set a change handler on a parent of input
elements, rather than setting one on each input element.

html5 also blesses this.  any standard dom onevent= can be attached
to any type of dom node.

so this fixes that by just adding all the supported onevent attributes
to the html whitelist.

this fixes https://code.google.com/p/google-caja/issues/detail?id=1531

R=metaweta@gmail.com