reading from /Users/felixz/caja/trunk/.appspot-change.framesec
fix es5 isDefinedInCajaFrame for firefox playground
http://codereview.appspot.com/6575043

ses has a function isDefinedInCajaFrame that works by iterating
over all the frames of the container page, then checking if
it's a caja frame, and seeing if the object in question inherits
from that frame's Object.

this assumes of course that you can examine objects in the frame,
but if the frame is in a different domain, like the plusone frame,
that's a security violation, which throws an exception in firefox,
and es5 dies.  (Chrome denies the access but doesn't throw an exception.)

this fixes that.

R=metaweta@gmail.com