fix es53 re.exec() for jquery 1.8 http://codereview.appspot.com/6567058
in the playground, running the jqueryui demos with jquery-1.8.1 fails,
the browser gets stuck in an infinite loop.

the reason is we're violating re.exec's contract.  re.exec returns
an array that has extra properties .index and .input, denoting
where and what got matched.  es53 doen't whitelist those properties
so when jquery tries to read match.index, it gets undefined, which
becomes NaN in arithmetic, which messes up the loop it's in.

this patch fixes that, and also updates jquery support to 1.8.2.

R=metaweta@gmail.com