add jquery-1.8.1 back to precajoled http://codereview.appspot.com/6560062
when I added jquery-1.8.2 to the precajole cache, I dropped 1.8.1,
but people are actually using 1.8.1, and it's kind of painful to
use any jquery without the precajoler (vs a few seconds extra
time building caja).  so I'm adding 1.8.1 back.

I have a change planned that will make the dynamic cajoler
cache much more effective, so we won't have to worry about this
as much, but I don't know yet when that will land.

R=metaweta@gmail.com