Distinguish between deferred scripts and scripts at the end of an HTML doc http://codereview.appspot.com/6485063
This removes the logic that assumes that all scripts at the end of
an HTML document are deferred scripts which should run after the
simulated call to document.close (HtmlEmitter.finish()) and
changes the RewriteHtmlStage which extracts scripts to embed an
explicit placeholder before the first deferred script.

It also modifies the RewriteHtmlStage to treat <script async> in
the same way as <script defer>.

Calls to document.write in deferred or async scripts have poorly
defined behavior so we prefer to have those fail.  Our taming of
document.write() will raise an exception if it is done after the
simulated document.close.

R=metaweta