Fix mutation-while-iterating bug in server side attribute sanitization. http://codereview.appspot.com/6640049
TemplateSanitizer.sanitizeAttrs did reverse-order iteration, which was
sufficient for when the sanitizer only deleted attributes but not now
that it adds virtualized attributes (r4783). This bug caused arbitrary
attributes to be skipped (left in the DOM) by the sanitizer.

Fixed by making a snapshot of the attribute list before any mutation.

This is not actually a vulnerability (that we have succeeded in
demonstrating), because the sanitizer happens to be always run twice
(in different pipeline stages), so the first run does all reordering
due to virtualization, and the second run catches the skipped attributes.

Additionally, this bug was exhibited in the TemplateSanitizerTest cases
when r5079 was implemented, but it was not recognized for what it was;
those test cases have now been corrected.

R=metaweta@gmail.com