Remove {height: 100%} (from r5021) and provide CSS class hooks instead. http://codereview.appspot.com/6496056
r5021 added height: 100% to the double div wrapper. While this allows
guest content to choose to fit the entire page, it also prevents guest
content from not taking up the entire page unless the host page happens
to provide an environment where there is no height to be 100% of.

Therefore, the explicit height is removed; in its place, CSS class names
are provided which allow the host page to arbitrarily style our wrapper
divs as they see fit.

R=jasvir@gmail.com