fix a.href fragment handling for jquery-ui tabs https://codereview.appspot.com/6815054/
r5080 changed how <a href='#foo'> behaves.  In particular,
a.hash used to be '#foo', but now it's '#foo-guest-suffix___',
which causes jquery-ui tabs to fail.

this CL fixes a.hash enough so that tabs work again.
there are potentially other problems with uri fragments,
and I stuck few TODOs for those.

R=jasvir, kpreid2