[APICHANGE] Fix reversal of caja-vdoc-* classes; add test for their presence. http://codereview.appspot.com/6784058
The class names caja-vdoc-inner and caja-vdoc-outer were applied in
reverse. Fix this, and add a test which confirms that the proper class
names are applied to the entire structure between the host's supplied
container node and the guest's root.

Impact: We have not instructed any user of Caja to make use of
caja-vdoc-inner or caja-vdoc-outer (only caja-vdoc-wrapper), but if
any do, they will need to reverse them.

R=ihab.awad@gmail.com