Improve debuggability of Rhino-driven test files. http://codereview.appspot.com/6827055
The test files which are executed by RhinoTestBed.runJsUnittestFromHtml
were written such that the URL / must be served off of ant-lib/, which
is incompatible with the test server started by 'ant brserve'. As a
workaround, the needed files from ant-lib are now copied to ant-testlib,
and the test HTML files now use relative URLs.

Also add all such test files to test-index.html.

R=metaweta@gmail.com