restrict form.submit http://codereview.appspot.com/6493115
currently, guest code can call form.submit() at any time.

this does not seem to be a serious problem, because those
submits usually trigger the pop-up blocker, but there
might still be potential for mischief.

this change restricts it so that the submit is ignored
unless it's executed during a user action, like in an
event handler responding to a mouse click, similar to
how we limit calls to element.focus().

this also narrows the restriction for focus() to just
click-like events, so you can't steal focus with, say,
a mousemove handler.

there's no automated test for this behavior yet, just a TODO,
because I tried several ways to write a purely html+js test
for it and failed miserably each time.

I did add a manual playground test, so it's not too hard to
check that it works as expected.

at some point, I'll write up some webdriver test logic for
this (and some other things that aren't being tested well).

R=metaweta@gmail.com