fix es53-test-domado-dom test failure in chrome http://codereview.appspot.com/6844068
es53-test-domado-dom fails in Chrome in es53 mode (but not es5 mode),
because Chrome webdriver can't click on some of the clickme elements,
because there's an absolute-position div overlapping the clickme elements.

The absolute-position div is nested inside a relative-position p
further down the page, so the div is intended to appear near the p.

The problem is es53 is strict about the html rule that p cannot
contain block elements, so the div gets moved outside the p.

So this fixes the failure by changing the enclosing p to a div.

I'm also fixing an inconsistent naming in testObjectClasses.
That's just a cosmetic issue; the test passes but doesn't turn
green because of the name mismatch.

R=kpreid2