quick fix for proxy cache problem http://codereview.appspot.com/6846111
For remote browser testing, the test urls were changed to use
http://some.ip.addr instead of http://localhost, but that
seems to not work if some.ip.addr is bounced through a proxy cache.

I'm going to fix our testing to do the appropriate no-cache stuff,
but this is a quick fix that changes it back to using localhost
unless you request a remote browser.

R=kpreid2