several fixes to browser testing speed https://codereview.appspot.com/6867084
This CL does a bunch of interrelated things, mostly about making
failing browser tests fail much faster.

1. Currently, a failing browser test will sit and wait until an 80s timeout,
because the driver scans for dom elements with the class "waiting".
Passing a test will remove the "waiting", but failing a test doesn't.

This CL makes failing tests also remove the "waiting", so that the test
driver will (usually) notice that the page has tried all the tests.

2. failed tests are now colored red (rather than remaining yellow).

3. The test driver loads a browser test page and waits up to N seconds for
all the tests on that page to finish.  This means the timeout has to be
large enough to accommodate pages with lots of tests, which means that
failure in a small test page might end up waiting much longer than
necessary for the timeout.

This CL adds a different timeout mechanism: rather than N seconds from
the start of the test, it times out N seconds after the last progress
change.

4. By default, webdriver.get() waits forever, so browser testing can
hang forever if something goes wrong while communicating with the
test server.

This CL adds calls to set webdriver timeouts to less than forever.

5. When a browser test is stuck, we timeout, mark it as an error,
and then try to open a new window for the next browser test, so
that the failing test's window can be examined later.  Trying to
open a new window in that situation can trigger the popup blocker,
and then the test driver hangs forever trying to switch to the
window that didn't get opened.

This CL checks if the window.open succeeded, and if it didn't,
it shrugs and continues testing.

6. Cajoler version skew tests are being tested by both the es5
driver and the es53 driver, which is silly since it always runs
in es53 mode.

7. caja.js version skew tests are being tested by both the es5
driver and the es53 driver, which is silly since this version
skew is supposed to throw an error early, at a point where
es5-vs-es53 isn't even relevant yet.

8. The test for cajoled minor version skew is wrong. It always
calls both fail and pass, and we only notice the pass not the fail.

This CL fixes that test to actually test what it's meant to tes.t

R=kpreid2