Remove traceStartup from Domado. https://codereview.appspot.com/6915044
traceStartup was left over from the original port of Domado, and has not
been used since, so I am removing it, to reduce minified code size and
the runtime cost of the no-op calls. I have replaced some of those which
were useful as signposts in the code with comments.

Also fixed a missing semicolon and added a note about named constructors.

R=metaweta@gmail.com