Memoize construction of tame NodeLists and friends. https://codereview.appspot.com/7060065
If cajaVM.makeArrayLike is a fully live implementation (due to the
browser supporting Proxies), then we can safely return the same
TameNodeList each time, so we now do for all NodeList-valued
properties.

Browsers also memoize NodeLists returned from getElementsByTagName
and getElementsByClassName, but not querySelectorAll, so we now do the
same.

The taming membrane is now made aware of node list tamings.

R=metaweta@gmail.com