Fixes issue 1622 -- top-level var mitigation fails because of bug in parser https://codereview.appspot.com/7099056
A bug in the upstream parser resulted in failure to mitigate gotchas if
a script contained a non-empty object literal eg. var bar = { foo : 1 }.

This change imports the updated fixed parser.

It also changes the behaviour such that if the parser fails, an error is
reported rather than proceeded without mitigating gotchas.

Added a unicode test and fixed the running of unicode tests under es5/3.

Added a FailureIsAnOption to es53 unicode tests because the caja.js api
does not currently expose a means to test failure to cajole.

R=kpreid.switchb.org@gmail.com