Fix another setTimeout-derived security bug. https://codereview.appspot.com/7001045
The guest-provided value of (virtual) window.onload was passed to
setTimeout without verification, allowing a string to be passed in and
thus evaluated in the host page context.

All occurrences of setTimeout with guest-provided callbacks in Domado
have now been funneled through the tamed setTimeout. Only window.onload
was specifically vulnerable, but I have changed the others to gain the
other robustness benefits outlined in tameSetAndClear.

R=metaweta@gmail.com