increase browser test timeout https://codereview.appspot.com/7220056
domado and other things have been getting slower, so it can take a
while for the browser to talk back to webdriver.  so this increases
the startup timeout, reducing the test flakiness.

this particular flakiness emits the helpful message "e is null" in
firefox's console.  that's a bug in webdriver failing to parse
something about firefox's stacktrace when it tries to serialize
the "timeout waiting for page to load" error back to the client.

this CL includes a small in-place patch of a js file in the webdriver
extension in the selenium jar, so that it doesn't throw "e is null",
and instead tries to report the error despite mismatched expectations.
(I'll try to send a fix back up to webdriver, this is just a hacky
short-term mitigation.)

R=kpreid2